generic template v0/requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
httpx==0.27.2
scikit-learn==1.5.2
numpy==2.1.1
joblib==1.4.2
